Restructure instantiator into modules
